+++ test/hooks/authorize/adapters/makeTests/remove.ts
@@ -1,0 +1,144 @@
+import assert from "assert";
+import feathers from "@feathersjs/feathers";
+import { createAliasResolver, defineAbility } from "@casl/ability";
+
+const resolveAction = createAliasResolver({
+  update: "patch",
+  read: ["get", "find"],
+  delete: "remove",
+});
+
+import { Application } from "@feathersjs/feathers";
+
+import authorize from "../../../../../lib/hooks/authorize/authorize.hook";
+import { Adapter, AuthorizeHookOptions } from "../../../../../lib/types";
+
+export default async function (
+  adapterName: Adapter,
+  makeService: () => unknown,
+  clean: (app, service) => Promise<void>,
+  authorizeHookOptions: Partial<AuthorizeHookOptions>,
+  afterHooks?: unknown[]
+): Promise<void> {
+  let app: Application;
+  let service;
+  let id;
+
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  const itSkip = (
+    adapterToTest: Adapter | Adapter[]
+  ): Mocha.TestFunction | Mocha.PendingTestFunction => {
+    const condition = (typeof adapterToTest === "string")
+      ? adapterName === adapterToTest
+      : adapterToTest.includes(adapterName);
+    return (condition)
+      ? it.skip
+      : it;
+  };
+      
+  describe(`${adapterName}: beforeAndAfter - remove:single`, function () {
+    beforeEach(async function () {
+      app = feathers();
+      app.use(
+        "tests",
+        makeService()
+      );
+      service = app.service("tests");
+  
+      // eslint-disable-next-line prefer-destructuring
+      id = service.options.id;
+  
+      const options = Object.assign({
+        availableFields: [id, "userId", "hi", "test", "published", "supersecret", "hidden"] 
+      }, authorizeHookOptions);
+      const allAfterHooks = [];
+      if (afterHooks) {
+        allAfterHooks.push(...afterHooks);
+      }
+      allAfterHooks.push(authorize(options));
+      //@ts-ignore
+      service.hooks({
+        before: {
+          all: [ authorize(options) ],
+        },
+        after: {
+          all: allAfterHooks
+        },
+      });
+  
+      await clean(app, service);
+    });
+    
+    it("can remove one item and return 'undefined' for not allowed read", async function () {
+      const item = await service.create({ test: true, userId: 1 });
+      
+      const updatedItem = await service.remove(item[id], {
+        //@ts-ignore
+        ability: defineAbility(can => {
+          can("remove", "tests");
+        }, { resolveAction })
+      });
+      
+      assert.deepStrictEqual(updatedItem, undefined, "removed item is undefined");
+      
+      const realItems = await service.find({ paginate: false }) as unknown[];
+      assert.strictEqual(realItems.length, 0, "no existent items");
+    });
+      
+    it("can remove one item and returns complete item", async function () {
+      const readMethods = ["read", "get"];
+      
+      for (const read of readMethods) {
+        await clean(app, service);
+        const item = await service.create({ test: true, userId: 1 });
+        const removedItem = await service.remove(item[id], {
+          //@ts-ignore
+          ability: defineAbility(can => {
+            can("remove", "tests");
+            can(read, "tests");
+          }, { resolveAction })
+        });
+        
+        assert.deepStrictEqual(removedItem, { [id]: item[id], test: true, userId: 1 }, `removed item correctly for read: '${read}'`);
+        
+        const realItems = await service.find({ paginate: false }) as unknown[];
+        assert.strictEqual(realItems.length, 0, "no existent items");
+      }
+    });
+      
+    it("throws if cannot remove item", async function () {
+      const item = await service.create({ test: true, userId: 1 });
+      
+      const promise = service.remove(item[id], {
+        //@ts-ignore
+        ability: defineAbility((can, cannot) => {
+          can("remove", "tests");
+          cannot("remove", "tests", { userId: 1 });
+        }, { resolveAction })
+      });
+      
+      assert.rejects(promise, err => err.name === "Forbidden", "cannot remove item");
+    });
+      
+    it("removes item and returns empty object for not overlapping '$select' and 'restricting fields'", async function() {
+      let item = { test: true, userId: 1, supersecret: true, hidden: true };
+      item = await service.create(item);
+      //@ts-ignore
+      const result = await service.remove(item[id], {
+        query: { $select: [id, "supersecret", "hidden"] },
+        //@ts-ignore
+        ability: defineAbility((can) => {
+          can("read", "tests", ["test", "userId"]);
+          can(["create", "remove"], "tests");
+        }, { resolveAction }),
+      });
+      assert.deepStrictEqual(result, {}, "returned item is empty because of $select and restricting fields");
+      await assert.rejects(
+        //@ts-ignore
+        service.get(item[id]),
+        err => err.name === "NotFound",
+        "item was deleted"
+      );
+    });
+  });
+}
