+++ test/hooks/authorize/adapters/mongodb.test.ts
@@ -2,9 +2,16 @@ import { MongoMemoryServer } from "mongodb-memory-server";
 import { MongoClient } from "mongodb";
 import { Service } from "feathers-mongodb";
 
-import makeTests from "./_makeTests";
+import makeTests from "./makeTests";
+import { ServiceCaslOptions } from "../../../../lib/types";
 
 let Model;
+
+declare module "@feathersjs/adapter-commons" {
+  interface ServiceOptions {
+    casl: ServiceCaslOptions
+  }
+}
 
 const makeService = () => {
   return new Service({
