+++ test/hooks/authorize/adapters/nedb.test.ts
@@ -1,13 +1,20 @@
 import NeDB from "nedb";
 import { Service } from "feathers-nedb";
-import makeTests from "./_makeTests";
+import makeTests from "./makeTests";
 import path from "path";
+import { ServiceCaslOptions } from "../../../../lib/types";
 
 // Create a NeDB instance
 const Model = new NeDB({
   filename: path.join(__dirname, "../../../.data/tests.db"),
   autoload: true
 });
+
+declare module "@feathersjs/adapter-commons" {
+  interface ServiceOptions {
+    casl: ServiceCaslOptions
+  }
+}
 
 const makeService = () => {
   return new Service({
