+++ test/hooks/authorize/adapters/objection.test.ts
@@ -1,9 +1,11 @@
 import { Model } from "objection";
-import makeTests from "./_makeTests";
+import makeTests from "./makeTests";
 import { Service } from "feathers-objection";
 import { getItems } from "feathers-hooks-common";
 import knex from "knex";
 import path from "path";
+import { ServiceCaslOptions } from "../../../../lib/types";
+import { HookContext } from "@feathersjs/feathers";
 
 const db  = knex({
   client: "sqlite3",
@@ -15,6 +17,12 @@ const db  = knex({
 });
 
 Model.knex(db);
+
+declare module "@feathersjs/adapter-commons" {
+  interface ServiceOptions {
+    casl: ServiceCaslOptions
+  }
+}
 
 class TestModel extends Model {
   static get tableName() {
@@ -64,7 +72,7 @@ const boolFields = [
 ];
 
 const afterHooks = [
-  context => {
+  (context: HookContext) => {
     let items = getItems(context);
     const isArray = Array.isArray(items);
     items = (isArray) ? items : [items];
